Reorganizes a little.
